Fixed error with duplicate pet search
- deleted duplicate and ensured proper one is updated
- worked a bit on refactoring code with app.js
