+++ kennel_link/app.js
@@ -4,7 +4,7 @@ const mongoose = require("./database");
 const client = require('./scripts/clientModel')
 const Employee = require('./db_modules/employee')
 const pet = require('./scripts/petModel')
-const Transaction = require('./db_modules/transaction')
+const Transaction = require('./scripts/transactionModel')
 const session = require('express-session')
 const reservation = require("./scripts/reservationModel")
 const signup_login = require("./scripts/signupLoginModel")
@@ -189,31 +189,9 @@ app.get("/transactions", (req,res) => {
   
 });
 
-// app.get("/add_pets", (req,res) => {
-//   const user = req.session.user;
-//   const user_type = req.session.type;
-//   res.render("pages/add_pets", {user: user, type: user_type});
-// })
-
-// app.get("/emp_pets", (req,res) => {
-//   res.render("pages/emp_pets");
-// })
-
-// app.get("/emp_transactions", (req,res) => {
-//   res.render("pages/emp_transactions")
-// })
-
 app.get("/emp_employees", (req,res) => {
   res.render("pages/emp_employees")
 })
-
-// app.get("/emp_clients_search", (req,res) => {
-  //   res.render("pages/emp_clients_search")
-// })
-
-// app.get("/emp_clients_edit", (req,res) => {
-//   res.render("pages/emp_clients_edit")
-// })
 
 app.get("/emp_reservation_add", (req,res) => {
   res.render("pages/emp_res_add")
@@ -238,46 +216,6 @@ app.get("/emp_transactions_search", (req,res) => {
 app.get("/emp_transactions_edit", (req,res) => {
   res.render("pages/emp_transactions_edit")
 })
-
-// app.get("/emp_clients_search", async (req,res) => {
-//   try {
-//     const result = await getClients(1, 5);
-//     // res.render("pages/emp_clients_search", {clients: result})
-//     // console.log(result);
-//     res.render("pages/emp_clients_search", {clients: result})
-//   } catch (error) {
-//     res.status(500).json({message: error.message});
-//     res.redirect('/emp_clients')
-//   }
-// })
-
-async function getNextID() {
-    try {
-        // Find the maximum employee ID
-        const maxEmployee = await Employee.find().sort({ empID: -1 }).limit(1);
-
-        // Find the maximum client ID
-        const maxClient = await Client.find().sort({ clientID: -1 }).limit(1);
-
-        // Determine the maximum ID from both collections
-        const maxID = Math.max((maxClient[0]?.clientID || 0), (maxEmployee[0]?.empID || 0)) + 1;
-
-        return maxID;
-    } catch (error) {
-        console.error("Error calculating next empID:", error);
-        throw error;
-    }
-}
-
-async function getClients(start, end) {
-  try {
-    const clients = await Client.find({activeFlag: true}).sort({clientLN:1}).skip(start-1).limit(end);
-    return clients;
-  } catch(error) {
-      console.error("Error returning client information:", error);
-      throw error;
-  }
-}
 
 let currentPage = 1;
 const pageSize = 10;
@@ -318,13 +256,6 @@ app.get("/emp_clients_edit", async (req, res) => {
   }
 });
 
-pet.getPets(1, 5)
-  .then(pets => {
-    console.log("Fetched pets");
-  })
-  .catch(error => {
-    console.error("Error fetching pets:", error);
-  });
 // Delete client
 app.post("/delete_client/:clientID", async (req, res) => {
   const clientID = req.params.clientID;
@@ -368,28 +299,6 @@ app.post("/update_client/:clientID", async (req, res) => {
   }
 });
 
-async function getPets(start, end, user_type, owner_id) {
-  if(user_type=='Client') {
-    try {
-      const pet_records = await Pet.find({ownerID: owner_id, activeFlag: true}).sort({petName:1}).skip(start-1).limit(end);
-      return pet_records;
-    } catch(error) {
-        console.error("Error returning client information:", error);
-        throw error;
-    }
-  }
-  if(user_type=='Employee') {
-    try {
-      const pet_records = await Pet.find({activeFlag: true}).sort({petName:1}).skip(start-1).limit(end);
-      return pet_records;
-    } catch(error) {
-        console.error("Error returning client information:", error);
-        throw error;
-    }
-  }
-
-};
-
 let currentPage_pets = 1;
 const pageSize_pets = 10;
 
@@ -398,7 +307,7 @@ app.get("/pets_search", async (req,res) => {
 
   if(user_type == 'Employee') {
     try {
-    const pet_records = await getPets(currentPage_pets, pageSize_pets, user_type, '');
+    const pet_records = await pet.getPets(currentPage_pets, pageSize_pets, user_type, '');
     const ownerIDs = pet_records.map(pet => pet.ownerID);
     const pet_owners = await Client.find({ clientID: { $in: ownerIDs } });
     const pet_owners_name = pet_owners.map(pet_owner => `${pet_owner.clientFN} ${pet_owner.clientLN}`);
@@ -411,7 +320,7 @@ app.get("/pets_search", async (req,res) => {
   }
   else if(user_type == 'Client') {
     const owner_id = parseInt(req.session.user.clientID);
-    const pet_records = await getPets(currentPage_pets, pageSize_pets, user_type, owner_id);
+    const pet_records = await pet.getPets(currentPage_pets, pageSize_pets, user_type, owner_id);
     res.render("pages/pets_search", { pets: pet_records, currentPage_pets, type: user_type});
 
   }
@@ -430,23 +339,10 @@ app.get("/pets_search/previous", async (req, res) => {
   res.redirect('/pets_search');
 });
 
-async function getPetById(pet_id) {
-  try {
-    const pet_record = await Pet.findOne({petID: pet_id});
-    if (!pet_record) {
-      console.log("petID undefined");
-    }
-    return pet_record;
-  } catch(error) {
-      console.error("Error returning client information:", error);
-      throw error;
-  }
-}
-
 app.get("/pets_edit", async (req, res) => {
   const petId = req.query.petId;
   try {
-    const found_pet = await getPetById(parseInt(petId)); // Fetch the client data
+    const found_pet = await pet.getPetById(parseInt(petId)); // Fetch the client data
     if (!found_pet) {
       return res.status(404).send("Pet not found");
     }
@@ -513,28 +409,9 @@ app.get("/emp_pets_search", async (req,res) => {
     }
     // Render the page with the fetched pets
     res.render("pages/emp_pets_search", { pets, searchQuery });
+  } catch(error) {
 
-async function getTrans(start, end, user_type, client_id) {
-  if(user_type=='Client') {
-    try {
-      const trans_records = await Transaction.find({clientID: client_id, activeFlag: true}).sort({transactionDate: -1}).skip(start-1).limit(end);
-      return trans_records;
-    } catch(error) {
-        console.error("Error returning transaction information:", error);
-        throw error;
-    }
-  }
-  if(user_type=='Employee') {
-    try {
-      const trans_records = await Transaction.find({activeFlag: true}).sort({transactionDate: -1}).skip(start-1).limit(end);
-      return trans_records;
-    } catch(error) {
-        console.error("Error returning transactioin information:", error);
-        throw error;
-    }
-  }
-
-};
+  }});
 
 let currentPage_trans = 1;
 const pageSize_trans = 10;
@@ -583,23 +460,11 @@ app.get("/transactions_search/previous", async (req, res) => {
   res.redirect('/transactions_search');
 });
 
-async function getTranById(trans_id) {
-  try {
-    const trans_record = await Transaction.findOne({TID: trans_id});
-    if (!trans_record) {
-      console.log("Transaction ID undefined");
-    }
-    return trans_record;
-  } catch(error) {
-      console.error("Error returning transaction information:", error);
-      throw error;
-  }
-}
 
 app.get("/transactions_edit", async (req, res) => {
   const transId = req.query.TID;
   try {
-    const found_trans = await getTranById(parseInt(transId)); // Fetch the client data
+    const found_trans = await Transaction.getTranById(parseInt(transId)); // Fetch the client data
     const clientName = req.query.client;
     if (!found_trans) {
       return res.status(404).send("Transaction not found");
